+++ modules/handle/src/cpp/usedHandle.cpp
@@ -33,7 +33,7 @@ usedHandle(const std::wstring& category)
         Dimensions dims(1, nbHandles);
         nelson_handle* nh
             = static_cast<nelson_handle*>(ArrayOf::allocateArrayOf(NLS_HANDLE, nbHandles));
-        for (int k = 0; k < nbHandles; k++) {
+        for (size_t k = 0; k < nbHandles; k++) {
             nh[k] = used[k];
         }
         res = ArrayOf(NLS_HANDLE, dims, (void*)nh);
